backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-core
langchain-openai
pypdf
python-dotenv
chromadb